Book and Author page changes
I tried to maintain your thematic styles using Lidarr/Radarr/Sonarr as examples, as well as the structure of the file. Almost all the fixes are for the author/book page, the rest flows from Lidarr fairly seamlessly. I did add color into the body element and add important, copying it from Radarr's base file. Otherwise, I wasn't sure how to get the "No History" for a book or author to be themed.