Duplicacy fixes
Fix the active tab indicator for the mobile sidebar, remove transparency on the pinned table header in restore window, *try* to fix the initial state of the "OK" button when trying to mark a job as parallel in the schedule window.

Placeholder text is not being differentiated from normal text in input boxes, so that's still pending along with the Schedule Graph navpills. Rest seems to look amazing to me now (up to you if you want to try to handle the multiple layers of cards on Schedule/Backup tabs, though I think it still looks good).